move (GIS / navigation) file format definitions to shared Django app #406
